backend/api/serializers.py
# serializers.py
from rest_framework import serializers
from .models import Attraction, Hotel

class AttractionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Attraction
        fields = '__all__'

class HotelSerializer(serializers.ModelSerializer):
    class Meta:
        model = Hotel
        fields = '__all__'
