requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
holidays==0.29
idna==3.2
joblib==1.1.0
numpy==1.21.2
packaging==21.2
pandas==1.3.3
platformdirs==2.4.0
pooch==1.5.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2021.3
redis==3.5.3
requests==2.26.0
scipy==1.7.3
shapely==1.7.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
typing-extensions>=3.10.0.2
tzdata==2021.5
urllib3==1.26.7
xgboost==1.7.6
